localizes card mocking fixture so CLI test modules will propery run.
